% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptsmooth.R
\name{ptsmooth}
\alias{ptsmooth}
\title{Smoothed Depth by Time Plot}
\usage{
ptsmooth(
  .dt,
  .x,
  .y,
  .val,
  .res_x = 0.5,
  .res_y = 2,
  y_grow_grid = TRUE,
  y_with_zero = TRUE
)
}
\arguments{
\item{.dt}{Data frame containing data to plot. Can be NULL, if all data
vectors are found in the enclosing environment.}

\item{.x}{Value that defines the x (time) coordinate of the plot.  Usually
a date or other time coordinate. Can be either a  vector or the name
of a vector in source data frame, not an expression.}

\item{.y}{Value that defines the (reversed) y coordinate of the plot.
Usually the depth.  Can be either a numeric vector or the name
of a numeric vector in source data frame, not an expression.}

\item{.val}{Value to be symbolized via the color of the dots.  Usually a
measured environmental variable that varies with water depth. Can be
either a numeric vector or the name of a numeric vector in source data
frame, not an expression.}

\item{.res_x}{Resolution for the plot, in the x, or time dimension.}

\item{.res_y}{Resolution for the plot in the y or depth dimension.}

\item{y_grow_grid}{TRUE/FALSE should the grid be expanded to integer depths?}

\item{y_with_zero}{TRUE/FALSE should the grid include depth = 0?}
}
\value{
A \code{ggplot} object (S3 class defined by \code{ggplot2})
}
\description{
Interpolates from a series of depth profiles, taken at different dates or
times, and generates a smoothed bivariate time (x axis) by depth (y axis,
reversed) plot using linear interpolation. Values of the variable being
plotted are symbolized by color.
}
\examples{
data(dep_sonde)
ptsmooth(dep_sonde, sample_date, depth, temp,
         .res_x = 0.5, .res_y = 5,
         y_grow_grid = TRUE,
         y_with_zero = TRUE)
}
