% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdggraph-package.R
\docType{package}
\name{tdggraph-package}
\alias{tdggraph}
\alias{tdggraph-package}
\title{tdggraph: Simple ggplot2 graphics for depth profiles in limnology and oceanography}
\description{
Limnology and oceanography frequently collect data along vertical
    profiles from the water's surface to depth. `tdggraph` provides functions
    built on `ggplot2` to produce graphics that handle profile data sets, 
    especially depth profile data collected repeatedly at the same location.
}
\author{
\strong{Maintainer}: Curtis Bohlen \email{curtis.bohlen@maine.edu}

}
\keyword{internal}
